artifacts() is only called from configs
